requirements.txt
asgiref==3.3.1
beautifulsoup4==4.8.2
boto3==1.17.43
botocore==1.20.43
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1.7
django-filter==2.4.0
django-modelcluster==5.1
django-sendmail-backend==0.1.2
django-storages==1.11.1
django-taggit==1.3.0
django-treebeard==4.5.1
djangorestframework==3.12.4
draftjs-exporter==2.1.7
et-xmlfile==1.0.1
flake8==3.9.0
google-api-core==1.26.3
google-api-python-client==2.1.0
google-auth==1.28.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
gunicorn==20.0.4
html5lib==1.1
httplib2==0.19.1
idna==2.10
jmespath==0.10.0
l18n==2020.6.1
mccabe==0.6.1
oauthlib==3.1.0
openpyxl==3.0.7
packaging==20.9
phonenumbers==8.12.20
Pillow==7.2.0
protobuf==3.15.6
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pyflakes==2.3.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rollbar==0.15.2
rsa==4.7.2
s3transfer==0.3.6
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
static3==0.7.0
tablib==3.0.0
Unidecode==1.2.0
uritemplate==3.0.1
urllib3==1.26.4
wagtail==2.10.2
webencodings==0.5.1
whitenoise==5.2.0
Willow==1.4
xlrd==2.0.1
XlsxWriter==1.3.8
xlwt==1.3.0
